backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
pydantic==2.9.0
pydantic-settings==2.5.2

# Transcription
openai-whisper==20240930
openai==1.51.0

# LLM Providers
anthropic==0.34.2
groq==0.11.0
httpx==0.27.2

# Audio processing
ffmpeg-python==0.2.0
pydub==0.25.1

# Utils
python-dotenv==1.0.1
aiofiles==24.1.0
structlog==24.4.0
